Attempting to set up and configure CORS
Utilising Django-CORS-Headers to get the hang of CORS (Cross-Origin Resource Sharing). Since we are likely will end up with an application where the React front end and the Django's APIs and backend components will be hosted on different ports, or protocols.
This is a common practice.

For example: If React was running on localhost:3000, then Django would be tied to localhost:8000. This separation would simplify the development process as two different developer teams would be able to work separately on front-end and back-end.